+++ tests/test_examples/test_quickstart.py
@@ -2,8 +2,8 @@
 import pytest
 
 from regress.ext.pytest import PytestContext, register_addoption
-from regress.fixture import RegressFixture
-from regress.pub import DefaultRegress
+from regress.regress import Regress
+from regress.pub import DefaultBaseRegress
 
 
 def pytest_addoption(parser):
@@ -12,22 +12,22 @@ def pytest_addoption(parser):
 
 @pytest.fixture(scope='module')
 def regress_instance():
-    regress = DefaultRegress()
-    regress.ensure_exists()
+    regress = DefaultBaseRegress()
+    regress.storage.ensure_exists()
     yield regress
 
 
 @pytest.fixture(scope='function')
 def regress(regress_instance, request):
-    fixture = RegressFixture(regress_instance,
-        PytestContext(request))
+    fixture = Regress(regress_instance,
+                      PytestContext(request))
     yield fixture
 # endregion: header
 
 @pytest.mark.no_canonize  # Here for not including to docs page
 # region: test_simple
 # regress is the helper fixture easy to setup
-def test_simple(regress: RegressFixture):
+def test_simple(regress: Regress):
     result = {'a': 1}
     regress.test(result)  # Commit first time
